Refactor LoggingUtils to move out codes related to logback
